added scrollIntoView
this is the simplest way to ensure that the highlighted note always remains in view. it's not the most optimal (since the note is often at the very right edge of the screen, or the very left edge of the screen, rather than in the center/middle), but it works. it's a good start.